server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.13.1
h11==0.14.0
identify==2.5.31
idna==3.4
isort==5.12.0
motor==3.3.2
nodeenv==1.8.0
passlib==1.7.4
platformdirs==3.11.0
pre-commit==3.5.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
setuptools==68.2.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
virtualenv==20.24.6
